% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEI_GSEA_gene_list.R
\name{DEI_GSEA_gene_list}
\alias{DEI_GSEA_gene_list}
\title{Retrieve a named numeric vector of the relative daily expression integral in log2}
\usage{
DEI_GSEA_gene_list(cond_1 = "SD", cond_2 = "LD", threshold = 0, NCBI_ID = TRUE)
}
\arguments{
\item{cond_1}{Either "EQ", "LD", or "SD"}

\item{cond_2}{Either "EQ", "LD", or "SD"}

\item{threshold}{Numeric. Filter for total expression level}

\item{NCBI_ID}{Logical. Whether the names should be converted from TAIR ID to NCBI ID}
}
\value{
Named numeric vector
}
\description{
Retrieve a named numeric vector of the relative daily expression integral in log2
}
