% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DEI_table}
\alias{DEI_table}
\title{Date frame with DEI data}
\format{
An object of class \code{data.frame} with 24845 rows and 4 columns.
}
\usage{
DEI_table
}
\description{
@format Data frame
}
\details{
\describe{
 \item{ID}{Gene identifier}
 \item{DEI_EQ}{Daily expression integral in equinox photoperiod}
 \item{DEI_LD}{Daily expression integral in long day photoperiod}
 \item{DEI_SD}{Daily expression integral in short day photoperiod}
}
}
\keyword{datasets}
