% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_GO.R
\name{heatmap_GO}
\alias{heatmap_GO}
\title{Make heatmap for GO enrichment}
\usage{
heatmap_GO(df, p_adj_threshold)
}
\arguments{
\item{df}{A list of gene identifiers in TAIR gene identifiers}

\item{p_adj_threshold}{Numeric. Filter for adjusted p value}
}
\value{
Named numeric vector
}
\description{
Make heatmap for GO enrichment
}
