% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{photoperiodic_clusters}
\alias{photoperiodic_clusters}
\title{List of photoperiodic clusters}
\format{
List of list
}
\usage{
photoperiodic_clusters
}
\description{
List of photoperiodic clusters
}
\keyword{datasets}
