% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_KEGGenrich.R
\name{run_KEGGenrich}
\alias{run_KEGGenrich}
\title{Perform KEGG term enrichment}
\usage{
run_KEGGenrich(geneList)
}
\arguments{
\item{geneList}{A list of gene identifiers in TAIR gene identifiers}
}
\value{
Named numeric vector
}
\description{
Perform KEGG term enrichment
}
